refactor: consolidate rules and treats into a Points page
Rules (how points are earned) and treats (how they're spent) are two
halves of the same economy, but treats were buried in Settings between
destructive maintenance actions. The Rules tab becomes Points, with
Earning and Treats sections; Settings keeps only data maintenance
(start date reset, recalculate).
