+++ apps/web/src/App.tsx
@@ -1,14 +1,14 @@
 import { Link, Route, Routes, useLocation } from "react-router-dom";
 import { AppShell, Button, Container, Group, Title } from "@mantine/core";
 import Dashboard from "./pages/Dashboard";
-import Rules from "./pages/Rules";
+import Points from "./pages/Points";
 import Activities from "./pages/Activities";
 import Settings from "./pages/Settings";
 
 const tabs = [
   { to: "/", label: "Dashboard" },
   { to: "/activities", label: "Activities" },
-  { to: "/rules", label: "Rules" },
+  { to: "/points", label: "Points" },
   { to: "/settings", label: "Settings" },
 ];
 
@@ -53,7 +53,7 @@ export default function App() {
           <Routes>
             <Route path="/" element={<Dashboard />} />
             <Route path="/activities" element={<Activities />} />
-            <Route path="/rules" element={<Rules />} />
+            <Route path="/points" element={<Points />} />
             <Route path="/settings" element={<Settings />} />
           </Routes>
         </Container>
